+++ lib/types/TcpSocket.d.ts
@@ -3,6 +3,10 @@
  *
  * @typedef {import('react-native').NativeEventEmitter} NativeEventEmitter
  *
+ * @typedef {{address: string, family: string, port: number}} AddressInfo
+ *
+ * @typedef {{localAddress: string, localPort: number, remoteAddress: string, remotePort: number, remoteFamily: string}} NativeConnectionInfo
+ *
  * @typedef {{
  * port: number;
  * host?: string;
@@ -15,20 +19,22 @@
  * tlsCheckValidity?: boolean,
  * tlsCert?: any,
  * }} ConnectionOptions
+ *
+ * @extends {EventEmitter<'connect' | 'timeout' | 'data' | 'error' | 'close', any>}
  */
-export default class TcpSocket extends EventEmitter {
+export default class TcpSocket extends EventEmitter<"connect" | "timeout" | "data" | "error" | "close", any> {
     /**
      * Initialices a TcpSocket.
      *
      * @param {number} id
      * @param {NativeEventEmitter} eventEmitter
-     * @param {string} [address]
+     * @param {NativeConnectionInfo} [connectionInfo]
      */
-    constructor(id: number, eventEmitter: NativeEventEmitter, address?: string | undefined);
-    /** @protected */
-    protected _id: number;
-    /** @protected */
-    protected _eventEmitter: import("react-native").NativeEventEmitter;
+    constructor(id: number, eventEmitter: NativeEventEmitter, connectionInfo?: NativeConnectionInfo | undefined);
+    /** @private */
+    private _id;
+    /** @private */
+    private _eventEmitter;
     /** @type {number} @private */
     private _timeoutMsecs;
     /** @private */
@@ -37,23 +43,16 @@ export default class TcpSocket extends EventEmitter {
     private _state;
     /** @private */
     private _encoding;
-    /**
-     * @protected
-     */
-    protected _registerEvents(): void;
-    _dataListener: import("react-native").EmitterSubscription | undefined;
-    _errorListener: import("react-native").EmitterSubscription | undefined;
-    _closeListener: import("react-native").EmitterSubscription | undefined;
-    _connectListener: import("react-native").EmitterSubscription | undefined;
-    /**
-     * @protected
-     */
-    protected _unregisterEvents(): void;
+    localAddress: string | undefined;
+    localPort: number | undefined;
+    remoteAddress: string | undefined;
+    remotePort: number | undefined;
+    remoteFamily: string | undefined;
     /**
      * @param {ConnectionOptions} options
-     * @param {(address: string) => void} [callback]
+     * @param {() => void} [callback]
      */
-    connect(options: ConnectionOptions, callback?: ((address: string) => void) | undefined): TcpSocket;
+    connect(options: ConnectionOptions, callback?: (() => void) | undefined): TcpSocket;
     _destroyed: boolean | undefined;
     /**
      * Sets the socket to timeout after `timeout` milliseconds of inactivity on the socket. By default `TcpSocket` do not have a timeout.
@@ -107,26 +106,19 @@ export default class TcpSocket extends EventEmitter {
      * @param {number} initialDelay ***IGNORED**. Default: `0`
      */
     setKeepAlive(enable?: boolean, initialDelay?: number): TcpSocket;
-    address(): string | undefined;
+    /**
+     * Returns the bound `address`, the address `family` name and `port` of the socket as reported
+     * by the operating system: `{ port: 12346, family: 'IPv4', address: '127.0.0.1' }`.
+     *
+     * @returns {AddressInfo | {}}
+     */
+    address(): AddressInfo | {};
     /**
      * @param {string | Buffer | Uint8Array} data
      * @param {BufferEncoding} [encoding]
      */
     end(data: string | Buffer | Uint8Array, encoding?: "ascii" | "utf8" | "utf-8" | "utf16le" | "ucs2" | "ucs-2" | "base64" | "latin1" | "binary" | "hex" | undefined): void;
     destroy(): void;
-    /**
-     * @private
-     * @param {string} address
-     */
-    private _onConnect;
-    /**
-     * @private
-     */
-    private _onClose;
-    /**
-     * @private
-     */
-    private _onError;
     /**
      * Sends data on the socket. The second parameter specifies the encoding in the case of a string — it defaults to UTF8 encoding.
      *
@@ -137,6 +129,20 @@ export default class TcpSocket extends EventEmitter {
      * @param {(error: string | null) => void} [callback]
      */
     write(buffer: string | Buffer | Uint8Array, encoding?: "ascii" | "utf8" | "utf-8" | "utf16le" | "ucs2" | "ucs-2" | "base64" | "latin1" | "binary" | "hex" | undefined, callback?: ((error: string | null) => void) | undefined): void;
+    ref(): void;
+    unref(): void;
+    /**
+     * @private
+     */
+    private _registerEvents;
+    _dataListener: import("react-native").EmitterSubscription | undefined;
+    _errorListener: import("react-native").EmitterSubscription | undefined;
+    _closeListener: import("react-native").EmitterSubscription | undefined;
+    _connectListener: import("react-native").EmitterSubscription | undefined;
+    /**
+     * @private
+     */
+    private _unregisterEvents;
     /**
      * @private
      * @param {string | Buffer | Uint8Array} buffer
@@ -145,19 +151,28 @@ export default class TcpSocket extends EventEmitter {
     private _generateSendBuffer;
     /**
      * @private
-     * @param {string} address
+     * @param {NativeConnectionInfo} connectionInfo
      */
     private _setConnected;
-    _address: string | undefined;
     /**
      * @private
      */
     private _setDisconnected;
-    ref(): void;
-    unref(): void;
 }
 export type BufferEncoding = "ascii" | "utf8" | "utf-8" | "utf16le" | "ucs2" | "ucs-2" | "base64" | "latin1" | "binary" | "hex";
 export type NativeEventEmitter = import("react-native").NativeEventEmitter;
+export type AddressInfo = {
+    address: string;
+    family: string;
+    port: number;
+};
+export type NativeConnectionInfo = {
+    localAddress: string;
+    localPort: number;
+    remoteAddress: string;
+    remotePort: number;
+    remoteFamily: string;
+};
 export type ConnectionOptions = {
     port: number;
     host?: string | undefined;
@@ -170,5 +185,5 @@ export type ConnectionOptions = {
     tlsCheckValidity?: boolean | undefined;
     tlsCert?: any;
 };
-import { EventEmitter } from "events";
+import EventEmitter from "eventemitter3";
 import { Buffer } from "buffer";